pyproject.toml
[project]
name = "goit-pythonweb-hw-08"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "fastapi (>=0.121.2,<0.122.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "pydantic[email] (>=2.12.4,<3.0.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "greenlet (>=3.2.4,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "fastapi-mail (<2.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "cloudinary (>=1.44.1,<2.0.0)",
    "redis (>=7.1.0,<8.0.0)",
    "libgravatar (>=1.0.4,<2.0.0)",
    "slowapi (>=0.1.9,<0.2.0)",
    "redis-lru (>=0.1.2,<0.2.0)",
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "aiosqlite (>=0.21.0,<0.22.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "alembic (>=1.17.2,<2.0.0)",
    "factory-boy (>=3.3.3,<4.0.0)",
    "testcontainers (>=4.13.3,<5.0.0)",
    "mypy (>=1.19.0,<2.0.0)",
    "ruff (>=0.14.8,<0.15.0)",
    "pytest (>=9.0.2,<10.0.0)",
    "anyio (>=4.12.0,<5.0.0)",
    "coverage (>=7.13.0,<8.0.0)",
    "types-passlib (>=1.7.7.20250602,<2.0.0.0)",
    "pytest-cov (>=7.0.0,<8.0.0)",
    "sphinx (>=9.0.4,<10.0.0)"
]

[tool.pytest.ini_options]
pythonpath = ["."]
